+++ src/lifeblood/net_messages/impl/message_protocol.py
