+++ src/lifeblood/net_messages/impl/tcp_simple_command_message_processor.py
