+++ src/lifeblood/scheduler_message_processor.py
@@ -9,11 +9,11 @@ from . import invocationjob
 from .exceptions import AlreadyRunning
 from .enums import WorkerPingReply, TaskScheduleStatus, WorkerState, WorkerType
 from .net_classes import WorkerResources
-from .net_messages.tcp_impl.tcp_simple_command_message_processor import TcpCommandMessageProcessor
-from .net_messages.tcp_impl.clients import CommandJsonMessageClient
+from .net_messages.impl.tcp_simple_command_message_processor import TcpCommandMessageProcessor
+from .net_messages.impl.clients import CommandJsonMessageClient
 from .net_messages.address import AddressChain
 from .net_messages.messages import Message
-from .net_messages.tcp_impl.message_haldlers import CommandMessageHandlerBase
+from .net_messages.impl.message_haldlers import CommandMessageHandlerBase
 
 
 from typing import Awaitable, Callable, Dict, Optional, Tuple, TYPE_CHECKING
