+++ src/lifeblood/worker_messsage_processor.py
@@ -8,11 +8,11 @@ from . import logging
 from . import invocationjob
 from .exceptions import AlreadyRunning
 from .enums import WorkerPingReply, TaskScheduleStatus
-from .net_messages.tcp_impl.tcp_simple_command_message_processor import TcpCommandMessageProcessor
-from .net_messages.tcp_impl.clients import CommandJsonMessageClient
+from .net_messages.impl.tcp_simple_command_message_processor import TcpCommandMessageProcessor
+from .net_messages.impl.clients import CommandJsonMessageClient
 from .net_messages.address import AddressChain
 from .net_messages.messages import Message
-from .net_messages.tcp_impl.message_haldlers import CommandMessageHandlerBase
+from .net_messages.impl.message_haldlers import CommandMessageHandlerBase
 
 
 from typing import Optional, Tuple, TYPE_CHECKING
